requirements.txt
absl-py==2.1.0
accelerate==0.31.0
aiohttp==3.9.5
aiosignal==1.3.1
albucore==0.0.13
albumentations==1.4.14
altair==6.0.0
annotated-types==0.6.0
anyio==4.12.1
asgiref==3.8.1
astunparse==1.6.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.2.0
blinker==1.9.0
blis==0.7.11
build==1.2.1
cachetools==5.5.0
catalogue==2.0.10
certifi==2023.5.7
cffi==2.0.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.5
click==8.1.7
cloudpathlib==0.18.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.0
cryptography==46.0.5
cycler==0.12.1
cymem==2.0.8
dataclasses-json==0.6.7
datasets==2.20.0
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dlib==19.24.2
eval_type_backport==0.2.0
evaluate==0.4.2
face-recognition==1.3.0
face-recognition-models==0.3.0
fastapi==0.110.0
filelock==3.13.1
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.49.0
fr-core-news-md==3.7.0
frozenlist==1.4.1
fsspec==2024.2.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-genai==1.65.0
google-generativeai==0.8.6
google-pasta==0.2.0
google-search-results==2.4.2
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.78.0
grpcio-status==1.71.2
h11==0.14.0
h5py==3.11.0
httpcore==1.0.5
httplib2==0.31.2
httptools==0.6.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.23.3
humanfriendly==10.0
hyperchain==0.1.3
idna==3.6
imageio==2.35.0
importlib_metadata==8.0.0
importlib_resources==6.4.3
inexactsearch==1.0.2
install==1.3.5
Jinja2==3.1.3
jiter==0.5.0
joblib==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
keras==3.5.0
kiwisolver==1.4.5
kubernetes==30.1.0
langchain==1.2.10
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.16
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langcodes==3.4.0
langgraph==1.0.10
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.9
langsmith==0.7.9
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mmh3==4.1.0
monotonic==1.6
monsterapi==1.0.9.2
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
narwhals==2.17.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.19.0
openai==1.41.0
opencv-python-headless==4.10.0.84
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
opt-einsum==3.3.0
optree==0.12.1
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==23.2
pandas==2.2.2
Pillow==10.1.0
pip==26.0.1
posthog==3.5.2
preshed==3.0.9
proto-plus==1.27.1
protobuf==5.29.6
psutil==5.9.8
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.13.1
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.1
pypdf==6.7.4
PyPika==0.48.9
pyproject_hooks==1.1.0
pyspellchecker==0.8.1
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.37.0
regex==2024.5.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rpds-py==0.30.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.2.2
scipy==1.10.1
sentence-transformers==3.0.1
serpapi==0.1.5
setuptools==65.5.0
shellingham==1.5.4
silpa-common==0.3
six==1.16.0
smart-open==7.0.4
smmap==5.0.2
sniffio==1.3.1
soundex==1.1.3
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spellchecker==0.4
SQLAlchemy==2.0.32
srsly==2.4.8
starlette==0.36.3
streamlit==1.54.0
sympy==1.12
ta==0.11.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.2
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.1.0
tifffile==2024.8.10
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.2.1
torchvision==0.17.1
tornado==6.5.4
tqdm==4.66.4
transformers==4.41.2
typer==0.12.4
typing_extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.2
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.2.1
uuid_utils==0.14.1
uvicorn==0.27.1
uvloop==0.20.0
wasabi==1.1.3
watchfiles==0.23.0
weasel==0.4.1
websocket-client==1.8.0
websockets==13.0
Werkzeug==3.0.3
wheel==0.44.0
wrapt==1.16.0
xxhash==3.6.0
yarl==1.9.4
zipp==3.20.0
zstandard==0.25.0